backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.31
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
python-multipart>=0.0.9
pydantic[email-validator]>=2.9.0
aiofiles>=24.1.0
slowapi>=0.1.9
